refactor: changed AppTheme folder into theme within naming convention
