+++ lib/app/app.locator.dart
@@ -10,7 +10,7 @@ import 'package:stacked/stacked.dart';
 import 'package:stacked/stacked_annotations.dart';
 import 'package:stacked_services/stacked_services.dart';
 
-import '../AppTheme/AppStateNotifier.dart';
+import '../theme/AppStateNotifier.dart';
 import '../services/functional_services/api_service.dart';
 import '../services/functional_services/authentication_service.dart';
 import '../services/functional_services/disk_space_service.dart';
