+++ lib/ui/widgets/dumb_widgets/frontpageview_appbar_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:rutorrentflutter/AppTheme/AppStateNotifier.dart';
+import 'package:rutorrentflutter/theme/AppStateNotifier.dart';
 import 'package:rutorrentflutter/models/account.dart';
 import 'package:rutorrentflutter/ui/views/home/home_viewmodel.dart';
