+++ lib/ui/widgets/smart_widgets/RSS Label Tile/RSS Detail Sheet/rss_detail_sheet_view.dart
@@ -1,5 +1,5 @@
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:rutorrentflutter/AppTheme/AppStateNotifier.dart';
+import 'package:rutorrentflutter/theme/AppStateNotifier.dart';
 import 'package:rutorrentflutter/models/rss.dart';
 import 'package:rutorrentflutter/ui/widgets/dumb_widgets/loading_shimmer.dart';
 import 'package:rutorrentflutter/ui/widgets/smart_widgets/RSS%20Label%20Tile/RSS%20Detail%20Sheet/rss_detail_sheet_viewmodel.dart';
